requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile --resolver=backtracking
#
bokeh==2.4.3
    # via -r requirements.in
click==8.1.3
    # via yake
colorama==0.4.6
    # via
    #   click
    #   tqdm
colorcet==3.0.1
    # via -r requirements.in
cython==0.29.32
    # via hdbscan
hdbscan==0.8.29
    # via -r requirements.in
jellyfish==0.9.0
    # via yake
jinja2==3.1.2
    # via bokeh
joblib==1.2.0
    # via
    #   hdbscan
    #   pynndescent
    #   scikit-learn
llvmlite==0.40.1
    # via
    #   numba
    #   pynndescent
markupsafe==2.1.1
    # via jinja2
networkx==2.8.8
    # via yake
numba==0.57.1
    # via
    #   pynndescent
    #   umap-learn
numpy==1.23.5
    # via
    #   bokeh
    #   hdbscan
    #   numba
    #   pandas
    #   pyarrow
    #   scikit-learn
    #   scipy
    #   umap-learn
    #   yake
packaging==21.3
    # via bokeh
pandas==2.0.2
    # via -r requirements.in
param==1.13.0
    # via pyct
pillow==9.3.0
    # via bokeh
pyarrow==12.0.1
    # via -r requirements.in
pyct==0.5.0
    # via colorcet
pynndescent==0.5.10
    # via umap-learn
pyparsing==3.0.9
    # via packaging
python-dateutil==2.8.2
    # via pandas
pytz==2022.6
    # via pandas
pyyaml==6.0
    # via bokeh
regex==2022.10.31
    # via segtok
scikit-learn==1.1.3
    # via
    #   -r requirements.in
    #   hdbscan
    #   pynndescent
    #   umap-learn
scipy==1.9.3
    # via
    #   hdbscan
    #   pynndescent
    #   scikit-learn
    #   umap-learn
segtok==1.5.11
    # via yake
six==1.16.0
    # via python-dateutil
tabulate==0.9.0
    # via yake
threadpoolctl==3.1.0
    # via scikit-learn
tornado==6.1
    # via bokeh
tqdm==4.65.0
    # via umap-learn
typing-extensions==4.4.0
    # via bokeh
tzdata==2023.3
    # via pandas
umap-learn==0.5.3
    # via -r requirements.in
yake==0.4.8
    # via -r requirements.in
